# Remove intermediate buffer in ASCII string read.