--- a/instrumentation/grpc/src/main/java/brave/grpc/TagContextBinaryMarshaller.java
+++ b/instrumentation/grpc/src/main/java/brave/grpc/TagContextBinaryMarshaller.java
@@ -2,6 +2,7 @@ package brave.grpc;
 
 import brave.internal.Platform;
 import io.grpc.Metadata.BinaryMarshaller;
+import java.nio.charset.Charset;
 import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.Map;
@@ -15,6 +16,8 @@ import java.util.Map;
 final class TagContextBinaryMarshaller implements BinaryMarshaller<Map<String, String>> {
   static final byte VERSION = 0, TAG_FIELD_ID = 0;
   static final byte[] EMPTY_BYTES = {};
+
+  static final Charset US_ASCII = Charset.forName("US-ASCII");
 
   @Override
   public byte[] toBytes(Map<String, String> tagContext) {
@@ -137,11 +140,10 @@ final class TagContextBinaryMarshaller implements BinaryMarshaller<Map<String, S
 
     String readAsciiString(int length) {
       if (length == -1 || remaining() < length) return null;
-      char[] string = new char[length];
-      for (int i = 0; i < length; i++) {
-        string[i] = (char) buf[pos++];
-      }
-      return new String(string);
+
+      String result = new String(buf, pos, length, US_ASCII);
+      pos += length;
+      return result;
     }
   }
 }